Add abstract class Component
