+++ logprep/abc/__init__.py
@@ -1,3 +1,6 @@
 # pylint: disable=missing-docstring
+from .component import Component
 from .processor import Processor
 from .connector import Connector
+from .input import Input
+from .output import Output
